\name{FastPCS-package}
\docType{package}
\title{Code to compute the FastPCS multivariate outlyingness index.}
\description{
Uses the FastPCS algorithm to compute the PCS multivariate 
	outlyingness index.
}
\details{
\tabular{ll}{
Package: \tab FastPCS\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2013-01-13\cr
Suggests: \tab mvtnorm\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
FastPCS                 Function to compute the FastPCS multivariate outlyingness
                       index.
NumStarts                 Internal function used to compute the FastPCS multivariate
                       outlyingness index.
quanf                     Internal function used to compute the FastPCS multivariate
                       outlyingness index.
}
}
\author{
Kaveh Vakili [aut, cre],
Maintainer: Kaveh Vakili <kaveh.vakili@wis.kuleuven.be>
}
\keyword{package}
